Gestion de quelques erreurs et améliorations

+++ controllers/sauce.js
@@ -91,7 +91,7 @@ exports.modifySauce = (req, res, next) => {
                     const sauceObject = JSON.parse(req.body.sauce);
                     delete sauceObject._id;
                     delete sauceObject._userId;
-                    delImg=true;
+                    delImg = true;
                     sauceUpdated = new Sauce({
                         ...sauceObject,
                         userId: req.auth.userId,
@@ -197,22 +197,23 @@ exports.deleteSauce = (req, res, next) => {
     Sauce.findOne({
         _id: req.params.id
     }).then(sauce => {
+        if (sauce.userId === req.auth.userId) {
+            deleteSauceImage(sauce.imageUrl);
 
-        deleteSauceImage(sauce.imageUrl);
-
-        Sauce.deleteOne({ _id: req.params.id }).then(
-            () => {
-                res.status(200).json({
-                    message: 'Deleted!'
-                });
-            }
-        ).catch(
-            (error) => {
-                res.status(400).json({
-                    error: error
-                });
-            }
-        );
+            Sauce.deleteOne({ _id: req.params.id }).then(
+                () => {
+                    res.status(200).json({
+                        message: 'Deleted!'
+                    });
+                }
+            ).catch(
+                (error) => {
+                    res.status(400).json({
+                        error: error
+                    });
+                }
+            );
+        }
     }
     ).catch(
         (error) => {
